refactor fmcs, initial commit with core module and ProcessingStrategy and ProcessingObject
